requirements.txt
torch==2.7.1
torchvision==0.22.1
torchaudio==2.7.1
cuequivariance-ops-torch-cu12==0.8.1
cuequivariance-torch==0.8.1
scipy>=1.9.0
ml_collections==1.1.0
tqdm==4.67.1
pandas==2.3.1
dm-tree==0.1.9
PyYAML==6.0.2
matplotlib==3.10.5
ipywidgets==8.1.7
py3Dmol==2.5.2
rdkit==2023.9.6
biopython==1.85
biotite==1.4.0
modelcif==1.4
gemmi==0.6.7
pdbeccdutils==0.8.6
fair-esm==2.0.0
scikit-learn==1.7.1
scikit-learn-extra==0.3.0
deepspeed==0.17.5  # Updated from 0.17.4 - part of fix for #182
pydantic>=2.0.0  # Explicitly specify v2.x for DeepSpeed compatibility - part of fix for #182
triton==3.3.1
optree==0.17.0
protobuf==6.31.1
icecream==2.1.7
ipdb==0.13.13
wandb==0.21.1
posix_ipc==1.3.0
numpy==1.26.4
